+++ backend/route/appointmentRoutes.js
@@ -1,28 +1,14 @@
 const express = require("express");
 const router = express.Router();
-const appointmentCtrl = require("../../controller/receptionist/appointmentController");
-const {
-  validateAppointment,
-} = require("../../validation/receptionist/appointmentValidator");
-const { runValidation } = require("../../validation/receptionist/Validate");
+const appointmentCtrl = require("../controller/appointmentController");
 
 // Appointment Endpoints
-router.post(
-  "/",
-  validateAppointment,
-  runValidation,
-  appointmentCtrl.createAppointment
-);
+router.post("/", appointmentCtrl.createAppointment);
 router.get("/", appointmentCtrl.getAppointmentsByDate);
 router.get("/patient/:patientId", appointmentCtrl.getAppointmentsByPatient);
 router.get("/doctor/:doctorId", appointmentCtrl.getAppointmentsByDoctor);
 router.get("/status", appointmentCtrl.getAppointmentsByStatus);
-router.put(
-  "/:appointmentId",
-  validateAppointment,
-  runValidation,
-  appointmentCtrl.updateAppointment
-);
+router.put("/:appointmentId", appointmentCtrl.updateAppointment);
 router.get("/:appointmentId", appointmentCtrl.getAppointmentById);
 router.patch("/:appointmentId/cancel", appointmentCtrl.cancelAppointment);
