--- backend/validation/receptionist/Validate.js
@@ -1,9 +1,0 @@
-const { validationResult } = require("express-validator");
-
-exports.runValidation = (req, res, next) => {
-  const errors = validationResult(req);
-  if (!errors.isEmpty()) {
-    return res.status(400).json({ errors: errors.array() });
-  }
-  next();
-};
